% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPath.R
\name{myPath}
\alias{myPath}
\title{Returns the absolute path of the script being called}
\usage{
myPath()
}
\description{
This path can be used to make R scripts portable, and
less rigidly dependent on a single filestructure.
Might be UNIX specific?
}
\examples{
myPath() 
}
\keyword{path}
